+++ ImTK/UI/Element/Basic/TextElement_Generic.cs
@@ -103,7 +103,7 @@ namespace ImTK.UI
                 unsafe {
                     size = ImGui.CalcTextSize((byte*)m_textBuffer.Data, (byte*)null, false, constraint.AvailableWidth);
                 }
-                return size;
+                return new System.Numerics.Vector2(MathF.Ceiling(size.X), MathF.Ceiling(size.Y));
             }
             else
             {
@@ -111,7 +111,7 @@ namespace ImTK.UI
                 unsafe {
                     size = ImGui.CalcTextSize((byte*)m_textBuffer.Data);
                 }
-                return new System.Numerics.Vector2(size.X, ImGui.GetTextLineHeight());
+                return new System.Numerics.Vector2(MathF.Ceiling(size.X), MathF.Ceiling(ImGui.GetTextLineHeight()));
             }
         }
     }