+++ ImTK/UI/Style/VisualElementStyle.cs
@@ -214,7 +214,8 @@ namespace ImTK.UI
                      prop.key == VisualElement.StyleKey.Bottom.Hash ||
                      prop.key == VisualElement.StyleKey.Left.Hash ||
                      prop.key == VisualElement.StyleKey.Right.Hash ||
-                     prop.key == VisualElement.StyleKey.Display.Hash)
+                     prop.key == VisualElement.StyleKey.Display.Hash ||
+                     prop.key == VisualElement.StyleKey.Overflow.Hash)
             {
                 prop.category = StyleCategory.Layout;
             }